+++ app/src/main/res/layout/circle.xml
@@ -2,29 +2,38 @@
 
 
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:app="http://schemas.android.com/apk/res-auto"
-              android:orientation="horizontal"
-              android:id="@+id/linerlayoutSingle"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content">
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/linerlayoutSingle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
     <ImageView
+            android:id="@+id/circleImage"
             android:layout_width="154dp"
-            android:layout_height="107dp" app:srcCompat="@mipmap/ic_launcher_round" android:id="@+id/circleImage"
-            android:layout_weight="1"/>
+            android:layout_height="107dp"
+            android:layout_weight="1"
+            app:srcCompat="@mipmap/ic_launcher_round" />
+
     <LinearLayout
-            android:orientation="vertical"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" android:layout_weight="1">
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            android:orientation="vertical">
+
         <TextView
-                android:text="TextView"
-                android:textStyle="bold"
-                android:textSize="17dp"
+                android:id="@+id/bookTitle"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
                 android:paddingTop="3dp"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" android:id="@+id/bookTitle"/>
-        <TextView
                 android:text="TextView"
+                android:textSize="17dp"
+                android:textStyle="bold" />
+
+        <TextView
+                android:id="@+id/bookContent"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content" android:id="@+id/bookContent"/>
+                android:layout_height="wrap_content"
+                android:text="TextView" />
     </LinearLayout>
 </LinearLayout>
