+++ app/src/main/res/layout/fragment_like_circle.xml
@@ -3,7 +3,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".ui.likeCircleFragment">
+                                                   tools:context=".ui.like.likeCircleFragment">
     <androidx.recyclerview.widget.RecyclerView
             android:layout_width="match_parent"
             android:layout_height="match_parent" android:id="@+id/likeRecyclerView"/>
