+++ app/src/main/res/layout/fragment_search_botton.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.searchBottonFragment" android:id="@+id/frameLayout1">
+        tools:context=".ui.serch.searchBottonFragment" android:id="@+id/frameLayout1">
     <EditText
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
