+++ app/src/main/res/layout/fragment_web_view.xml
@@ -3,7 +3,7 @@
                                                    xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".ui.webViewFragment">
+                                                   tools:context=".ui.webview.webViewFragment">
     <WebView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
